fastdtw handle higher dim

+++ src/utils.jl
@@ -1,10 +1,10 @@
-Base.@propagate_inbounds Base.getindex(v::AbstractVector, ::typeof(!), i::Int) = v[i]
-Base.@propagate_inbounds Base.getindex(v::AbstractMatrix, ::typeof(!), i::Int) = @view v[:,i]
-Base.@propagate_inbounds Base.getindex(v::AbstractArray{<:Any,3}, ::typeof(!), i::Int) = @view v[:,:,i]
+Base.@propagate_inbounds Base.getindex(v::AbstractVector, ::typeof(!), i) = v[i]
+Base.@propagate_inbounds Base.getindex(v::AbstractMatrix, ::typeof(!), i) = @view v[:,i]
+Base.@propagate_inbounds Base.getindex(v::AbstractArray{<:Any,3}, ::typeof(!), i) = @view v[:,:,i]
 
-Base.@propagate_inbounds Base.setindex!(v::AbstractVector, val, ::typeof(!), i::Int) = v[i] = val
-Base.@propagate_inbounds Base.setindex!(v::AbstractMatrix, val, ::typeof(!), i::Int) = v[:,i] .= val
-Base.@propagate_inbounds Base.setindex!(v::AbstractArray{<:Any,3}, val, ::typeof(!), i::Int) = v[:,:,i] .= val
+Base.@propagate_inbounds Base.setindex!(v::AbstractVector, val, ::typeof(!), i) = v[i] = val
+Base.@propagate_inbounds Base.setindex!(v::AbstractMatrix, val, ::typeof(!), i) = v[:,i] .= val
+Base.@propagate_inbounds Base.setindex!(v::AbstractArray{<:Any,3}, val, ::typeof(!), i) = v[:,:,i] .= val
 
 lastlength(x) = size(x, ndims(x))
